Use AutomatchPreferences type for AutomatchSettings as well.
